Remove certificate logic from ModelUtils class.
