+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/CruiseControl.java
@@ -32,6 +32,7 @@ import io.strimzi.api.kafka.model.template.DeploymentTemplate;
 import io.strimzi.api.kafka.model.template.InternalServiceTemplate;
 import io.strimzi.api.kafka.model.template.PodTemplate;
 import io.strimzi.certs.CertAndKey;
+import io.strimzi.certs.SecretCertProvider;
 import io.strimzi.operator.cluster.ClusterOperatorConfig;
 import io.strimzi.operator.cluster.model.cruisecontrol.Capacity;
 import io.strimzi.operator.cluster.model.logging.LoggingModel;
@@ -471,8 +472,8 @@ public class CruiseControl extends AbstractModel implements SupportsMetrics, Sup
         }
         LOGGER.debugCr(reconciliation, "End generating certificates");
 
-        return ModelUtils.buildSecret(namespace, CruiseControlResources.secretName(cluster), ccCerts,
-                labels, ownerReference, clusterCa.caCertGenerationFullAnnotation(), Map.of());
+        return ModelUtils.createSecret(CruiseControlResources.secretName(cluster), namespace, labels, ownerReference,
+                SecretCertProvider.buildSecretData(ccCerts), clusterCa.caCertGenerationFullAnnotation(), Map.of());
     }
 
     /**
