+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/EntityUserOperator.java
@@ -265,7 +265,7 @@ public class EntityUserOperator extends AbstractModel implements SupportsLogging
      */
     public Secret generateSecret(ClusterCa clusterCa, boolean isMaintenanceTimeWindowsSatisfied) {
         Secret secret = clusterCa.entityUserOperatorSecret();
-        return ModelUtils.buildSecret(reconciliation, clusterCa, secret, namespace, KafkaResources.entityUserOperatorSecretName(cluster), componentName,
+        return clusterCa.buildTrustedCertificateSecret(secret, namespace, KafkaResources.entityUserOperatorSecretName(cluster), componentName,
             CERT_SECRET_KEY_NAME, labels, ownerReference, isMaintenanceTimeWindowsSatisfied);
     }
