+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/ZookeeperCluster.java
@@ -35,6 +35,7 @@ import io.strimzi.api.kafka.model.template.PodTemplate;
 import io.strimzi.api.kafka.model.template.ResourceTemplate;
 import io.strimzi.api.kafka.model.template.ZookeeperClusterTemplate;
 import io.strimzi.certs.CertAndKey;
+import io.strimzi.certs.SecretCertProvider;
 import io.strimzi.operator.cluster.model.jmx.JmxModel;
 import io.strimzi.operator.cluster.model.jmx.SupportsJmx;
 import io.strimzi.operator.cluster.model.logging.LoggingModel;
@@ -451,8 +452,8 @@ public class ZookeeperCluster extends AbstractModel implements SupportsMetrics, 
             throw new RuntimeException("Failed to prepare ZooKeeper certificates", e);
         }
 
-        return ModelUtils.buildSecret(namespace, KafkaResources.zookeeperSecretName(cluster), certs,
-                labels, ownerReference, clusterCa.caCertGenerationFullAnnotation(), emptyMap());
+        return ModelUtils.createSecret(KafkaResources.zookeeperSecretName(cluster), namespace, labels, ownerReference,
+                SecretCertProvider.buildSecretData(certs), clusterCa.caCertGenerationFullAnnotation(), emptyMap());
     }
 
     /* test */ Container createContainer(ImagePullPolicy imagePullPolicy) {
